% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_auditors.R
\name{chkargs}
\alias{chkargs}
\title{checks the default arguments == arguments passed via command line}
\usage{
chkargs(args, parser, addlog = TRUE)
}
\arguments{
\item{args}{named list of arguments as returned by the parser}

\item{parser}{parser object used to hold defaults and parse command line}
}
\description{
checks the default arguments == arguments passed via command line
}
