% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_auditors.R
\name{default_checker}
\alias{default_checker}
\title{checks the default arguments == arguments in python code}
\usage{
default_checker(args, python_code)
}
\arguments{
\item{args}{named list of arguments as returned by the parser}

\item{python_code}{code generated by argparse parser specifications}
}
\description{
checks the default arguments == arguments in python code
}
